GroundCheck Bug Fixed + DashAbility Implementation
Before Groundcheck was using the whole characters collision box to check if it was on the ground. Solution used was to add a child gameobject "Feet" and use that object's collision box for checking if the player was on the ground or not.
A physics asset was added to the player's collider box to ensure that they did not stick to the walls.
Also added a quick mapping of a platforming level just to test movement and finetune things.
Also added DashAbility as well as templates to add Melee and Shooting