Added code for shooting a projectile + BasicEnemyHealth the rest of the environemnt pngs
Modified how the player's model gets flipped. Instead of using spriterenderer().flipx now we are keeping track of it as a variable and rotating its transform. This allowed for easier manipulation of FirePoint, the position where the player shoots the projectile.
Currently, the projectile used is a basic sphere (placeholder). Code for enemy makes it so that once it gets hit by a projectile, it will lose HP.